[Silver III] Title: 2×n 타일링, Time: 168 ms, Memory: 17816 KB -BaekjoonHub

+++ 백준/Silver/11726. 2×n 타일링/2×n 타일링.java
@@ -3,7 +3,7 @@ public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
         int N = sc.nextInt();
-        long D[] = new long[1001];
+        int D[] = new int[1001];
         D[1] = 1;  // N = 1 일때 타일 채우는 경우의 수
         D[2] = 2;  // N = 2 일떼 타일 채우는 경우의 수
